Move InstallDropdownButton to its own file
