+++ www/package.json
@@ -53,7 +53,7 @@
     "grommet": "^2.24.0",
     "grommet-icons": "^4.7.0",
     "highlight.js": "^11.5.1",
-    "honorable": "^0.110.0",
+    "honorable": "^0.111.0",
     "honorable-theme-default": "^0.56.0",
     "js-yaml": "^4.1.0",
     "lodash": "^4.17.21",
