+++ www/src/components/repository/RepositoryHeader.js
@@ -1,6 +1,6 @@
 import { useContext, useState } from 'react'
 import { useMutation } from '@apollo/client'
-import { A, Button, Div, DropdownButton, ExtendTheme, Flex, H1, H2, H3, Icon, Img, MenuItem, Modal, P, Pre, Span } from 'honorable'
+import { A, Button, Div, Flex, H1, H3, Icon, Img, Modal, P, Pre, Span } from 'honorable'
 import { GitHubIcon, InstalledLabel, LinksIcon, Tag, TrashCanIcon } from 'pluralsh-design-system'
 
 import RepositoryContext from '../../contexts/RepositoryContext'
@@ -9,59 +9,10 @@ import usePaginatedQuery from '../../hooks/usePaginatedQuery'
 
 import { capitalize } from '../../utils/string'
 
-import Code from '../utils/Code'
-
 import { DELETE_INSTALLATION_MUTATION, RECIPES_QUERY } from './queries'
+import { providerToIcon, providerToIconHeight } from './constants'
 
-const providerToIcon = {
-  AWS: `${process.env.PUBLIC_URL}/aws-icon.png`,
-  AZURE: `${process.env.PUBLIC_URL}/azure.png`,
-  EQUINIX: `${process.env.PUBLIC_URL}/equinix-metal.png`,
-  GCP: `${process.env.PUBLIC_URL}/gcp.png`,
-  KIND: `${process.env.PUBLIC_URL}/kind.png`,
-}
-
-const providerToIconHeight = {
-  AWS: 12,
-  AZURE: 14,
-  EQUINIX: 16,
-  GCP: 14,
-  KIND: 14,
-}
-
-const providerToDisplayName = {
-  AWS: 'Amazon Web Services',
-  AZURE: 'Azure',
-  EQUINIX: 'Equinix Metal',
-  GCP: 'Google Cloud Platform',
-  KIND: 'Kind',
-}
-
-const extendedTheme = {
-  DropdownButton: {
-    Button: [
-      {
-        borderRadius: 'normal',
-      },
-    ],
-    Menu: [
-      {
-        width: 256 + 64 + 16 + 4,
-        left: 'unset',
-      },
-    ],
-  },
-  MenuItem: {
-    Root: [
-      {
-        borderBottom: '1px solid border',
-        '&:last-of-type': {
-          borderBottom: 'none',
-        },
-      },
-    ],
-  },
-}
+import InstallDropdownButton from './InstallDropdownButton'
 
 function InstalledActions({ installation, ...props }) {
   const [modalOpen, setModalOpen] = useState(false)
@@ -133,100 +84,6 @@ function InstalledActions({ installation, ...props }) {
             Delete
           </Button>
         </Flex>
-      </Modal>
-    </>
-  )
-}
-
-function InstallDropdownButton({ recipes, ...props }) {
-  const { name } = useContext(RepositoryContext)
-  const [recipe, setRecipe] = useState(null)
-
-  function renderModalContent() {
-    if (!recipe) return null
-
-    return (
-      <Div minWidth={512}>
-        <Flex
-          align="center"
-          as={H2}
-        >
-          Install {capitalize(name)} on {providerToDisplayName[recipe.provider]}
-          <Img
-            ml={0.75}
-            alt={recipe.name}
-            src={providerToIcon[recipe.provider]}
-            height={1.5 * providerToIconHeight[recipe.provider]}
-          />
-        </Flex>
-        <P mt={2}>
-          {capitalize(recipe.description)}.
-        </P>
-        <P mt={1}>
-          In your installation repository run:
-        </P>
-        <Code
-          language="bash"
-          mt={2}
-        >
-          {`plural bundle install ${name} ${recipe.name}`}
-        </Code>
-      </Div>
-    )
-  }
-
-  return (
-    <>
-      <ExtendTheme theme={extendedTheme}>
-        <DropdownButton
-          fade
-          label="Install"
-          onChange={event => setRecipe(event.target.value)}
-          {...props}
-        >
-          {recipes.map(recipe => (
-            <MenuItem
-              key={recipe.id}
-              value={recipe}
-            >
-              <Flex>
-                <Flex
-                  align="center"
-                  justify="center"
-                  width={2 * 16}
-                  flexShrink={0}
-                >
-                  <Img
-                    alt={recipe.name}
-                    src={providerToIcon[recipe.provider]}
-                    height={1.5 * providerToIconHeight[recipe.provider]}
-                  />
-                </Flex>
-                <Div
-                  ml={1}
-                  flexShrink={0}
-                  flexGrow={1}
-                  flexBasis="calc(100% - 4 * 16px)"
-                >
-                  <P fontWeight={500}>
-                    {providerToDisplayName[recipe.provider]}
-                  </P>
-                  <P
-                    wordBreak="break-word"
-                  >
-                    {capitalize(recipe.description)}
-                  </P>
-                </Div>
-              </Flex>
-            </MenuItem>
-          ))}
-        </DropdownButton>
-      </ExtendTheme>
-      <Modal
-        open={!!recipe}
-        onClose={() => setRecipe(null)}
-      >
-        {renderModalContent()}
       </Modal>
     </>
   )
