+++ www/src/components/repository/constants.js
@@ -1,0 +1,23 @@
+export const providerToIcon = {
+  AWS: `${process.env.PUBLIC_URL}/aws-icon.png`,
+  AZURE: `${process.env.PUBLIC_URL}/azure.png`,
+  EQUINIX: `${process.env.PUBLIC_URL}/equinix-metal.png`,
+  GCP: `${process.env.PUBLIC_URL}/gcp.png`,
+  KIND: `${process.env.PUBLIC_URL}/kind.png`,
+}
+
+export const providerToIconHeight = {
+  AWS: 12,
+  AZURE: 14,
+  EQUINIX: 16,
+  GCP: 14,
+  KIND: 14,
+}
+
+export const providerToDisplayName = {
+  AWS: 'Amazon Web Services',
+  AZURE: 'Azure',
+  EQUINIX: 'Equinix Metal',
+  GCP: 'Google Cloud Platform',
+  KIND: 'Kind',
+}
